Enhancements
- Added support for following additional domains
- Removed request callbacks
- Moved follow logic rules from global to per-request

+++ crawl_response.go
@@ -38,7 +38,7 @@ type CrawlResponse struct {
 	Request       *CrawlRequest
 	Duration      time.Duration
 	StatusCode    int
-	ContentLength int
+	ContentLength int64
 	ContentType   string
 	URLs          []string
 }
